Player got in, jump another player when going backward (error fixed)
When a player moves backward, if there is another player on the next token, the player can skip the token and move to the next next token until there is no player on it. If a player successfully comes back to the origin, he won't roll dice when he has turn. Instead, his next player rolls dice.